uv-client: fix linehaul test
This test was introduced in 42973cd9cb098e59f75e287d72a425af8a614e91. It
looks like it compares some values against some platform specific code
that attempts to find the OS version. But the comparisons made some
assumptions about what kind of data is available. In this commit, we try
to make the test a little more flexible on Linux by not assuming that
`Option` values are `Some`.