uv: tweak 'no_solution' test
Bumping EXCLUDE_NEWER seemed to change more in this test's output than
just version numbers. It looks like it depends on a constraint for the
most recently published version of Flask. So we just bump the version
number. Now the test output largely stays the same, with the exception
of changes in version numbers (which we expect).