uv/tests: add regression test for `echo pendulum | uv pip compile -`
I stumbled across this when writing tests for
`--emit-marker-expressions`. Namely, I observed that in CI, the `tzdata`
dependency of `pendulum` wasn't included in the `requirements.txt`
output on Windows.

@konstin [suggested] that this was a bug, so I've created a test for it.
In particular, it looks like [`tzdata` is an unconditional dependency of
`pendulum`][tzdata-unconditional].

It turns out that this is because of a [heuristic we use on
Windows][windows-filters] where `tzdata` is excluded from the output in
an effort to make most tests pass in a platform agnostic manner.

So this commit:

1. Adds a comment documenting why we're doing that.
2. Keeps the regression test but with Windows filters disabled to ensure
we're getting this particular case correct anyway. Because why not.

[suggested]: https://github.com/astral-sh/uv/pull/2651#discussion_r1539722464
[tzdata-unconditional]: https://inspector.pypi.io/project/pendulum/3.0.0/packages/67/5e/e646afbd1632bfbacdae79289d7d5879efdeeb5f5e58327bc5c698731107/pendulum-3.0.0-cp310-none-win_amd64.whl/pendulum-3.0.0.dist-info/METADATA#line.12
[windows-filters]: https://github.com/astral-sh/uv/blob/f9db0ed8f0fc50972f2a5d99f4099d6ca7d49f52/crates/uv/tests/common/mod.rs#L420-L445