HACK: uv-resolver: skip marker evaluation
This is a hack to skip marker evaluation in and around the resolver.

This doesn't apply to every single marker evaluation in `uv`. For
example, I think the marker evaluation that is done as part of `uv pip
install` is fine for now. (Although eventually I think that will need to
get refactored.)