Make fetching work when package is in Amazon S3
When a package is hosted on Amazon S3 (sometimes the case when using
Gemfury as a private repository), there might be a redirect from the
Gemfury link to the S3 link. The S3 link will be different for HEAD and
GET requests. For this reason, we need to use the original Gemfury link
when passing arguments to the range reader.