Improve operating system and architecture types in `uv-toolchain` (#2918)
Prompted in
https://github.com/astral-sh/uv/pull/2842#discussion_r1555640300

Updates the toolchain `Os` and `Arch` types to be a closer match to
those in `platform-tags`.

They don't match exactly, e.g. we track `Libc` separately instead of
using `Manylinux` and `Musllinux` and we don't need release versions.
This makes me think we might not actually want to consolidate to central
shared types.