Warn on invalid pyproject.toml
When we try to compile a pyproject.toml, we currently silently fall back to PEP 517 when we can't parse it (according to PEP 621). This PR adds a warning. It would be even better if we could be strict and require a valid pyproject.toml.

This is relevant for `tool.uv.sources`, which doesn't get forwarded through PEP 517, so we'd silently lose that information.